% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{read_selected_text}
\alias{read_selected_text}
\title{Read selected text}
\usage{
read_selected_text()
}
\description{
.. content for \description{} (Reads in text that is selected) ..
}
\details{
.. content for \details{} (Reads in text that is selected)..
}
